requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
playwright==1.44.0
python-telegram-bot==21.3
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.4
aiofiles==23.2.1